backend/requirements.txt
fastapi
uvicorn
pymongo
spacy
pdfminer.six
python-multipart
apscheduler
requests
beautifulsoup4
python-dotenv
PyJWT
Werkzeug
